TDX: fix tdx_allowed_port() and warn on disallow

+++ arch/x86/include/asm/tdx.h
@@ -44,7 +44,7 @@ phys_addr_t tdx_shared_mask(void);
 
 int tdx_hcall_request_gpa_type(phys_addr_t start, phys_addr_t end, bool enc);
 
-bool tdx_allowed_port(short int port);
+bool tdx_allowed_port(int port);
 
 int tdx_mcall_tdreport(u64 data, u64 reportdata);
